refactor(scripts): add converter
